Implement New Entity Action and tweaked the UI Metamodel

+++ opaeum-runtime/opaeum-runtime-domain/src/main/java/org/opaeum/runtime/domain/EnumResolver.java
@@ -2,7 +2,7 @@ package org.opaeum.runtime.domain;
 
 
 public interface EnumResolver{
-	int toOpaeumId(IEnum e);
-	IEnum fromOpaeumId(int i);
+	long toOpaeumId(IEnum e);
+	IEnum fromOpaeumId(long i);
 }
