+++ opaeum-runtime/opaeum-runtime-domain/src/main/java/org/opaeum/runtime/environment/marshall/EnumValue.java
@@ -3,11 +3,11 @@ package org.opaeum.runtime.environment.marshall;
 
 public class EnumValue extends Value{
 	private static final long serialVersionUID = 9086348287296846913L;
-	Integer enumId;
-	public Integer getEnumId(){
+	Long enumId;
+	public Long getEnumId(){
 		return enumId;
 	}
-	public EnumValue(String classId,Integer enumId){
+	public EnumValue(String classId,Long enumId){
 		super(classId);
 		this.enumId = enumId;
 	}
